// my-management-mate/src/app/enum/category.enum.ts
export enum ItemCategory {
    Stationery = "Stationery",
    OfficeEquipment = "OfficeEquipment",
    OfficeFurniture = "OfficeFurniture",
    OfficeSupplies = "OfficeSupplies",
    StorageOrganization = "StorageOrganization",
    SecurityPrivacy = "SecurityPrivacy",
    Miscellaneous = "Miscellaneous",
}